// src/app/shared/types/api-output-types/fleet-business/control-configuration-output.interface.ts
import { BidoEquipmentDTO } from '../../api-types/bido-equipment-dto.interface';
import { BidoFlDTO } from '../../api-types/bido-fl-dto.interface';
import { BidtStockMvtDTO } from '../../api-types/bidt-stock-mvt-dto.interface';
import { BireItemDTO } from '../../api-types/bire-item-dto.interface';

export interface ControlConfigurationOutput {
  bidoEquipment?: BidoEquipmentDTO;
  bidoFl?: BidoFlDTO;
  bireItem?: BireItemDTO;
  installedStockMvtData?: BidtStockMvtDTO;
  removedStockMvtData?: BidtStockMvtDTO;
  asset?: string;
  fatherAsset?: string;
  functionCode?: string;
  checkConfigurationCause?: string;
  checkConfigurationCauseAlt?: string;
  checkConfigurationResult?: string;
  checkCoreCause?: string;
  checkCoreCauseAlt?: string;
  checkCoreResult?: string;
  checkFlCause?: string;
  checkFlCauseAlt?: string;
  checkFlResult?: string;
  checkLogbookCause?: string;
  checkLogbookCauseAlt?: string;
  checkLogbookResult?: string;
  checkMelCause?: string;
  checkMelCauseAlt?: string;
  checkMelResult?: string;
  checkPotentialCause?: string;
  checkPotentialCauseAlt?: string;
  checkPotentialResult?: string;
  checkReferentialCause?: string;
  checkReferentialCauseAlt?: string;
  checkReferentialResult?: string;
  checkStructureCause?: string;
  checkStructureCauseAlt?: string;
  checkStructureResult?: string;
  isAssetOrFatherAssetInstalledDuringMaintenance?: boolean;
  isAssetRemovedDuringMaintenance?: boolean;
  itemMandatory?: boolean;
  replacedAsset?: string;
  replacedBidoEquipment?: BidoEquipmentDTO;
  structure?: string;
  uniqueBidoEquipmentChild?: BidoEquipmentDTO;
  workOrderId?: string;
  smallestRemainingInCyclesValue?: number;
  smallestRemainingInHoursValue?: number;
  closestCalendarLimitDate?: Date;
  children?: ControlConfigurationOutput[];
  equipmentAssetCount?: number;

  checkGlobalResult?: string;
}