// src/app/shared/types/api-types/bido-fl-dto.interface.ts
export interface BidoFlDTO {
  flCode?: string;
  flFlCode?: string;
  flDescription?: string;
  flFunction?: string;
  flItems?: string;
  flIsMel?: boolean;
  flNominalQuantity?: number;
  flRequiredQuantity?: number;
  flDelayCategory?: string;
  flComments?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  flIsOpeConf?: boolean;
  flOrderNumber?: number;
  flPositionX?: number;
  flPositionY?: number;
  flIsInConfSheet?: boolean;
}